minor restructures
- added port flag to allow multiple instances
- added query support via json payload and form as well as queryparams
- updated build file to work with new structure
- executable now otf-classifier
